+++ csd/api_client.go
@@ -22,23 +22,25 @@ type ApiClient struct {
 	UserAgent       string
 }
 
-type Zone struct {
+// Zone Delegation
+
+type ZoneDelegation struct {
 	Name        string   `json:"name"`
 	NameServers []string `json:"name_servers"`
 }
 
-func (c *ApiClient) createZone(zone Zone) (Zone, diag.Diagnostics) {
+func (c *ApiClient) createZoneDelegation(zoneDelegation ZoneDelegation) (ZoneDelegation, diag.Diagnostics) {
 	var diags diag.Diagnostics
 
 	buffer := new(bytes.Buffer)
-	if err := json.NewEncoder(buffer).Encode(zone); err != nil {
-		return zone, diag.FromErr(err)
+	if err := json.NewEncoder(buffer).Encode(zoneDelegation); err != nil {
+		return zoneDelegation, diag.FromErr(err)
 	}
 
 	client := &http.Client{Timeout: 10 * time.Second}
-	request, err := http.NewRequest(http.MethodPost, fmt.Sprintf("%s/v1/zones", HostURL), buffer)
+	request, err := http.NewRequest(http.MethodPost, fmt.Sprintf("%s/v2/zone_delegations", HostURL), buffer)
 	if err != nil {
-		return zone, diag.FromErr(err)
+		return zoneDelegation, diag.FromErr(err)
 	}
 	authorizationHeaders := signRequest(request, c.AccessKeyId, c.SecretAccessKey, c.SessionToken)
 	request.Header.Add("X-Amz-Security-Token", c.SessionToken)
@@ -50,7 +52,7 @@ func (c *ApiClient) createZone(zone Zone) (Zone, diag.Diagnostics) {
 
 	response, err := client.Do(request)
 	if err != nil {
-		return zone, diag.FromErr(err)
+		return zoneDelegation, diag.FromErr(err)
 	}
 	defer response.Body.Close()
 
@@ -58,9 +60,9 @@ func (c *ApiClient) createZone(zone Zone) (Zone, diag.Diagnostics) {
 		// Create proper error message if AWS credentials are not valid, probably because they expired
 		var responseBody map[string]string
 		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
-			return zone, diag.FromErr(err)
+			return zoneDelegation, diag.FromErr(err)
 		}
-		return zone, append(diags, diag.Diagnostic{
+		return zoneDelegation, append(diags, diag.Diagnostic{
 			Severity: diag.Error,
 			Summary:  "Couldn't authenticate to API, please check AWS credentials",
 			Detail:   responseBody["message"],
@@ -68,37 +70,37 @@ func (c *ApiClient) createZone(zone Zone) (Zone, diag.Diagnostics) {
 	} else if response.StatusCode == 409 {
 		var responseBody map[string]string
 		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
-			return zone, diag.FromErr(err)
+			return zoneDelegation, diag.FromErr(err)
 		}
-		return zone, append(diags, diag.Diagnostic{
+		return zoneDelegation, append(diags, diag.Diagnostic{
 			Severity: diag.Error,
-			Summary:  "Couldn't create zone",
+			Summary:  "Couldn't create zone delegation",
 			Detail:   responseBody["message"],
 		})
 	} else if response.StatusCode != 201 {
 		// Create error message for any other unexpected errors
 		body, _ := io.ReadAll(response.Body)
-		return zone, append(diags, diag.Diagnostic{
+		return zoneDelegation, append(diags, diag.Diagnostic{
 			Severity: diag.Error,
 			Summary:  "Unexpected error message from API",
 			Detail:   fmt.Sprintf("HTTP %d: %s", response.StatusCode, body),
 		})
 	}
 
-	if err := json.NewDecoder(response.Body).Decode(&zone); err != nil {
-		return zone, diag.FromErr(err)
+	if err := json.NewDecoder(response.Body).Decode(&zoneDelegation); err != nil {
+		return zoneDelegation, diag.FromErr(err)
 	}
-	return zone, diags
+	return zoneDelegation, diags
 }
 
-func (c *ApiClient) getZone(name string) (Zone, diag.Diagnostics) {
+func (c *ApiClient) getZoneDelegation(name string) (ZoneDelegation, diag.Diagnostics) {
 	var diags diag.Diagnostics
-	var zone Zone
+	var zoneDelegation ZoneDelegation
 
 	client := &http.Client{Timeout: 10 * time.Second}
-	request, err := http.NewRequest(http.MethodGet, fmt.Sprintf("%s/v1/zones/%s", HostURL, name), strings.NewReader(""))
+	request, err := http.NewRequest(http.MethodGet, fmt.Sprintf("%s/v2/zone_delegations/%s", HostURL, name), strings.NewReader(""))
 	if err != nil {
-		return zone, diag.FromErr(err)
+		return zoneDelegation, diag.FromErr(err)
 	}
 	authorizationHeaders := signRequest(request, c.AccessKeyId, c.SecretAccessKey, c.SessionToken)
 	request.Header.Add("X-Amz-Security-Token", c.SessionToken)
@@ -110,7 +112,7 @@ func (c *ApiClient) getZone(name string) (Zone, diag.Diagnostics) {
 
 	response, err := client.Do(request)
 	if err != nil {
-		return zone, diag.FromErr(err)
+		return zoneDelegation, diag.FromErr(err)
 	}
 	defer response.Body.Close()
 
@@ -118,9 +120,9 @@ func (c *ApiClient) getZone(name string) (Zone, diag.Diagnostics) {
 		// Create proper error message if AWS credentials are not valid, probably because they expired
 		var responseBody map[string]string
 		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
-			return zone, diag.FromErr(err)
+			return zoneDelegation, diag.FromErr(err)
 		}
-		return zone, append(diags, diag.Diagnostic{
+		return zoneDelegation, append(diags, diag.Diagnostic{
 			Severity: diag.Error,
 			Summary:  "Couldn't authenticate to API, please check AWS credentials",
 			Detail:   responseBody["message"],
@@ -128,37 +130,37 @@ func (c *ApiClient) getZone(name string) (Zone, diag.Diagnostics) {
 	} else if response.StatusCode == 404 {
 		var responseBody map[string]string
 		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
-			return zone, diag.FromErr(err)
+			return zoneDelegation, diag.FromErr(err)
 		}
-		return zone, append(diags, diag.Diagnostic{
+		return zoneDelegation, append(diags, diag.Diagnostic{
 			Severity: diag.Error,
-			Summary:  "Couldn't find zone with given name",
+			Summary:  "Couldn't find zone delegation with given name",
 			Detail:   responseBody["message"],
 		})
 	} else if response.StatusCode != 200 {
 		// Create error message for any other unexpected errors
 		body, _ := io.ReadAll(response.Body)
-		return zone, append(diags, diag.Diagnostic{
+		return zoneDelegation, append(diags, diag.Diagnostic{
 			Severity: diag.Error,
 			Summary:  "Unexpected error message from API",
 			Detail:   fmt.Sprintf("HTTP %d: %s", response.StatusCode, body),
 		})
 	}
 
-	if err := json.NewDecoder(response.Body).Decode(&zone); err != nil {
-		return zone, diag.FromErr(err)
+	if err := json.NewDecoder(response.Body).Decode(&zoneDelegation); err != nil {
+		return zoneDelegation, diag.FromErr(err)
 	}
-	return zone, diags
+	return zoneDelegation, diags
 }
 
-func (c *ApiClient) getZones() ([]Zone, diag.Diagnostics) {
+func (c *ApiClient) getZoneDelegations() ([]ZoneDelegation, diag.Diagnostics) {
 	var diags diag.Diagnostics
-	var zones []Zone
+	var zoneDelegations []ZoneDelegation
 
 	client := &http.Client{Timeout: 10 * time.Second}
-	request, err := http.NewRequest(http.MethodGet, fmt.Sprintf("%s/v1/zones", HostURL), strings.NewReader(""))
+	request, err := http.NewRequest(http.MethodGet, fmt.Sprintf("%s/v2/zone_delegations", HostURL), strings.NewReader(""))
 	if err != nil {
-		return zones, diag.FromErr(err)
+		return zoneDelegations, diag.FromErr(err)
 	}
 	authorizationHeaders := signRequest(request, c.AccessKeyId, c.SecretAccessKey, c.SessionToken)
 	request.Header.Add("X-Amz-Security-Token", c.SessionToken)
@@ -170,7 +172,7 @@ func (c *ApiClient) getZones() ([]Zone, diag.Diagnostics) {
 
 	response, err := client.Do(request)
 	if err != nil {
-		return zones, diag.FromErr(err)
+		return zoneDelegations, diag.FromErr(err)
 	}
 	defer response.Body.Close()
 
@@ -178,9 +180,9 @@ func (c *ApiClient) getZones() ([]Zone, diag.Diagnostics) {
 		// Create proper error message if AWS credentials are not valid, probably because they expired
 		var responseBody map[string]string
 		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
-			return zones, diag.FromErr(err)
+			return zoneDelegations, diag.FromErr(err)
 		}
-		return zones, append(diags, diag.Diagnostic{
+		return zoneDelegations, append(diags, diag.Diagnostic{
 			Severity: diag.Error,
 			Summary:  "Couldn't authenticate to API, please check AWS credentials",
 			Detail:   responseBody["message"],
@@ -188,31 +190,31 @@ func (c *ApiClient) getZones() ([]Zone, diag.Diagnostics) {
 	} else if response.StatusCode != 200 {
 		// Create error message for any other unexpected errors
 		body, _ := io.ReadAll(response.Body)
-		return zones, append(diags, diag.Diagnostic{
+		return zoneDelegations, append(diags, diag.Diagnostic{
 			Severity: diag.Error,
 			Summary:  "Unexpected error message from API",
 			Detail:   fmt.Sprintf("HTTP %d: %s", response.StatusCode, body),
 		})
 	}
 
-	if err := json.NewDecoder(response.Body).Decode(&zones); err != nil {
-		return zones, diag.FromErr(err)
+	if err := json.NewDecoder(response.Body).Decode(&zoneDelegations); err != nil {
+		return zoneDelegations, diag.FromErr(err)
 	}
-	return zones, diags
+	return zoneDelegations, diags
 }
 
-func (c *ApiClient) updateZone(zone Zone) (Zone, diag.Diagnostics) {
+func (c *ApiClient) updateZoneDelegation(zoneDelegation ZoneDelegation) (ZoneDelegation, diag.Diagnostics) {
 	var diags diag.Diagnostics
 
 	buffer := new(bytes.Buffer)
-	if err := json.NewEncoder(buffer).Encode(zone); err != nil {
-		return zone, diag.FromErr(err)
+	if err := json.NewEncoder(buffer).Encode(zoneDelegation); err != nil {
+		return zoneDelegation, diag.FromErr(err)
 	}
 
 	client := &http.Client{Timeout: 10 * time.Second}
-	request, err := http.NewRequest(http.MethodPut, fmt.Sprintf("%s/v1/zones/%s", HostURL, zone.Name), buffer)
+	request, err := http.NewRequest(http.MethodPut, fmt.Sprintf("%s/v2/zone_delegations/%s", HostURL, zoneDelegation.Name), buffer)
 	if err != nil {
-		return zone, diag.FromErr(err)
+		return zoneDelegation, diag.FromErr(err)
 	}
 	authorizationHeaders := signRequest(request, c.AccessKeyId, c.SecretAccessKey, c.SessionToken)
 	request.Header.Add("X-Amz-Security-Token", c.SessionToken)
@@ -224,7 +226,7 @@ func (c *ApiClient) updateZone(zone Zone) (Zone, diag.Diagnostics) {
 
 	response, err := client.Do(request)
 	if err != nil {
-		return zone, diag.FromErr(err)
+		return zoneDelegation, diag.FromErr(err)
 	}
 	defer response.Body.Close()
 
@@ -232,9 +234,9 @@ func (c *ApiClient) updateZone(zone Zone) (Zone, diag.Diagnostics) {
 		// Create proper error message if AWS credentials are not valid, probably because they expired
 		var responseBody map[string]string
 		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
-			return zone, diag.FromErr(err)
+			return zoneDelegation, diag.FromErr(err)
 		}
-		return zone, append(diags, diag.Diagnostic{
+		return zoneDelegation, append(diags, diag.Diagnostic{
 			Severity: diag.Error,
 			Summary:  "Couldn't authenticate to API, please check AWS credentials",
 			Detail:   responseBody["message"],
@@ -242,34 +244,34 @@ func (c *ApiClient) updateZone(zone Zone) (Zone, diag.Diagnostics) {
 	} else if response.StatusCode == 404 {
 		var responseBody map[string]string
 		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
-			return zone, diag.FromErr(err)
+			return zoneDelegation, diag.FromErr(err)
 		}
-		return zone, append(diags, diag.Diagnostic{
+		return zoneDelegation, append(diags, diag.Diagnostic{
 			Severity: diag.Error,
-			Summary:  "Couldn't find zone with given name",
+			Summary:  "Couldn't find zone delegation with given name",
 			Detail:   responseBody["message"],
 		})
 	} else if response.StatusCode != 200 {
 		// Create error message for any other unexpected errors
 		body, _ := io.ReadAll(response.Body)
-		return zone, append(diags, diag.Diagnostic{
+		return zoneDelegation, append(diags, diag.Diagnostic{
 			Severity: diag.Error,
 			Summary:  "Unexpected error message from API",
 			Detail:   fmt.Sprintf("HTTP %d: %s", response.StatusCode, body),
 		})
 	}
 
-	if err := json.NewDecoder(response.Body).Decode(&zone); err != nil {
-		return zone, diag.FromErr(err)
+	if err := json.NewDecoder(response.Body).Decode(&zoneDelegation); err != nil {
+		return zoneDelegation, diag.FromErr(err)
 	}
-	return zone, diags
+	return zoneDelegation, diags
 }
 
-func (c *ApiClient) deleteZone(name string) diag.Diagnostics {
+func (c *ApiClient) deleteZoneDelegation(name string) diag.Diagnostics {
 	var diags diag.Diagnostics
 
 	client := &http.Client{Timeout: 10 * time.Second}
-	request, err := http.NewRequest(http.MethodDelete, fmt.Sprintf("%s/v1/zones/%s", HostURL, name), strings.NewReader(""))
+	request, err := http.NewRequest(http.MethodDelete, fmt.Sprintf("%s/v2/zone_delegations/%s", HostURL, name), strings.NewReader(""))
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -305,7 +307,310 @@ func (c *ApiClient) deleteZone(name string) diag.Diagnostics {
 		}
 		return append(diags, diag.Diagnostic{
 			Severity: diag.Error,
-			Summary:  "Couldn't find zone with given name",
+			Summary:  "Couldn't find zone delegation with given name",
+			Detail:   responseBody["message"],
+		})
+	} else if response.StatusCode != 204 {
+		// Create error message for any other unexpected errors
+		body, _ := io.ReadAll(response.Body)
+		return append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Unexpected error message from API",
+			Detail:   fmt.Sprintf("HTTP %d: %s", response.StatusCode, body),
+		})
+	}
+
+	return diags
+}
+
+// Record
+
+type Record struct {
+	Name   string `json:"name"`
+	Value  string `json:"value"`
+	TTL    int    `json:"ttl"`
+	RRType string `json:"rrtype"`
+}
+
+func (c *ApiClient) createRecord(record Record) (Record, diag.Diagnostics) {
+	var diags diag.Diagnostics
+
+	buffer := new(bytes.Buffer)
+	if err := json.NewEncoder(buffer).Encode(record); err != nil {
+		return record, diag.FromErr(err)
+	}
+
+	client := &http.Client{Timeout: 10 * time.Second}
+	request, err := http.NewRequest(http.MethodPost, fmt.Sprintf("%s/v2/records", HostURL), buffer)
+	if err != nil {
+		return record, diag.FromErr(err)
+	}
+	authorizationHeaders := signRequest(request, c.AccessKeyId, c.SecretAccessKey, c.SessionToken)
+	request.Header.Add("X-Amz-Security-Token", c.SessionToken)
+	request.Header.Add("X-Amz-Date", authorizationHeaders.date)
+	request.Header.Add("Authorization", authorizationHeaders.authorizationHeaders)
+	request.Header.Add("content-type", "application/json")
+	request.Header.Add("x-amz-content-sha256", fmt.Sprintf("%x", authorizationHeaders.payloadHash))
+	request.Header.Set("User-Agent", c.UserAgent)
+
+	response, err := client.Do(request)
+	if err != nil {
+		return record, diag.FromErr(err)
+	}
+	defer response.Body.Close()
+
+	if response.StatusCode == 403 {
+		// Create proper error message if AWS credentials are not valid, probably because they expired
+		var responseBody map[string]string
+		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
+			return record, diag.FromErr(err)
+		}
+		return record, append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Couldn't authenticate to API, please check AWS credentials",
+			Detail:   responseBody["message"],
+		})
+	} else if response.StatusCode == 409 {
+		var responseBody map[string]string
+		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
+			return record, diag.FromErr(err)
+		}
+		return record, append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Couldn't create record",
+			Detail:   responseBody["message"],
+		})
+	} else if response.StatusCode != 201 {
+		// Create error message for any other unexpected errors
+		body, _ := io.ReadAll(response.Body)
+		return record, append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Unexpected error message from API",
+			Detail:   fmt.Sprintf("HTTP %d: %s", response.StatusCode, body),
+		})
+	}
+
+	if err := json.NewDecoder(response.Body).Decode(&record); err != nil {
+		return record, diag.FromErr(err)
+	}
+	return record, diags
+}
+
+func (c *ApiClient) getRecord(name string) (Record, diag.Diagnostics) {
+	var diags diag.Diagnostics
+	var record Record
+
+	client := &http.Client{Timeout: 10 * time.Second}
+	request, err := http.NewRequest(http.MethodGet, fmt.Sprintf("%s/v2/records/%s", HostURL, name), strings.NewReader(""))
+	if err != nil {
+		return record, diag.FromErr(err)
+	}
+	authorizationHeaders := signRequest(request, c.AccessKeyId, c.SecretAccessKey, c.SessionToken)
+	request.Header.Add("X-Amz-Security-Token", c.SessionToken)
+	request.Header.Add("X-Amz-Date", authorizationHeaders.date)
+	request.Header.Add("Authorization", authorizationHeaders.authorizationHeaders)
+	request.Header.Add("content-type", "application/json")
+	request.Header.Add("x-amz-content-sha256", fmt.Sprintf("%x", authorizationHeaders.payloadHash))
+	request.Header.Set("User-Agent", c.UserAgent)
+
+	response, err := client.Do(request)
+	if err != nil {
+		return record, diag.FromErr(err)
+	}
+	defer response.Body.Close()
+
+	if response.StatusCode == 403 {
+		// Create proper error message if AWS credentials are not valid, probably because they expired
+		var responseBody map[string]string
+		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
+			return record, diag.FromErr(err)
+		}
+		return record, append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Couldn't authenticate to API, please check AWS credentials",
+			Detail:   responseBody["message"],
+		})
+	} else if response.StatusCode == 404 {
+		var responseBody map[string]string
+		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
+			return record, diag.FromErr(err)
+		}
+		return record, append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Couldn't find record with given name",
+			Detail:   responseBody["message"],
+		})
+	} else if response.StatusCode != 200 {
+		// Create error message for any other unexpected errors
+		body, _ := io.ReadAll(response.Body)
+		return record, append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Unexpected error message from API",
+			Detail:   fmt.Sprintf("HTTP %d: %s", response.StatusCode, body),
+		})
+	}
+
+	if err := json.NewDecoder(response.Body).Decode(&record); err != nil {
+		return record, diag.FromErr(err)
+	}
+	return record, diags
+}
+
+func (c *ApiClient) getRecords() ([]Record, diag.Diagnostics) {
+	var diags diag.Diagnostics
+	var record []Record
+
+	client := &http.Client{Timeout: 10 * time.Second}
+	request, err := http.NewRequest(http.MethodGet, fmt.Sprintf("%s/v2/records", HostURL), strings.NewReader(""))
+	if err != nil {
+		return record, diag.FromErr(err)
+	}
+	authorizationHeaders := signRequest(request, c.AccessKeyId, c.SecretAccessKey, c.SessionToken)
+	request.Header.Add("X-Amz-Security-Token", c.SessionToken)
+	request.Header.Add("X-Amz-Date", authorizationHeaders.date)
+	request.Header.Add("Authorization", authorizationHeaders.authorizationHeaders)
+	request.Header.Add("content-type", "application/json")
+	request.Header.Add("x-amz-content-sha256", fmt.Sprintf("%x", authorizationHeaders.payloadHash))
+	request.Header.Set("User-Agent", c.UserAgent)
+
+	response, err := client.Do(request)
+	if err != nil {
+		return record, diag.FromErr(err)
+	}
+	defer response.Body.Close()
+
+	if response.StatusCode == 403 {
+		// Create proper error message if AWS credentials are not valid, probably because they expired
+		var responseBody map[string]string
+		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
+			return record, diag.FromErr(err)
+		}
+		return record, append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Couldn't authenticate to API, please check AWS credentials",
+			Detail:   responseBody["message"],
+		})
+	} else if response.StatusCode != 200 {
+		// Create error message for any other unexpected errors
+		body, _ := io.ReadAll(response.Body)
+		return record, append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Unexpected error message from API",
+			Detail:   fmt.Sprintf("HTTP %d: %s", response.StatusCode, body),
+		})
+	}
+
+	if err := json.NewDecoder(response.Body).Decode(&record); err != nil {
+		return record, diag.FromErr(err)
+	}
+	return record, diags
+}
+
+func (c *ApiClient) updateRecord(record Record) (Record, diag.Diagnostics) {
+	var diags diag.Diagnostics
+
+	buffer := new(bytes.Buffer)
+	if err := json.NewEncoder(buffer).Encode(record); err != nil {
+		return record, diag.FromErr(err)
+	}
+
+	client := &http.Client{Timeout: 10 * time.Second}
+	request, err := http.NewRequest(http.MethodPut, fmt.Sprintf("%s/v2/records/%s", HostURL, record.Name), buffer)
+	if err != nil {
+		return record, diag.FromErr(err)
+	}
+	authorizationHeaders := signRequest(request, c.AccessKeyId, c.SecretAccessKey, c.SessionToken)
+	request.Header.Add("X-Amz-Security-Token", c.SessionToken)
+	request.Header.Add("X-Amz-Date", authorizationHeaders.date)
+	request.Header.Add("Authorization", authorizationHeaders.authorizationHeaders)
+	request.Header.Add("content-type", "application/json")
+	request.Header.Add("x-amz-content-sha256", fmt.Sprintf("%x", authorizationHeaders.payloadHash))
+	request.Header.Set("User-Agent", c.UserAgent)
+
+	response, err := client.Do(request)
+	if err != nil {
+		return record, diag.FromErr(err)
+	}
+	defer response.Body.Close()
+
+	if response.StatusCode == 403 {
+		// Create proper error message if AWS credentials are not valid, probably because they expired
+		var responseBody map[string]string
+		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
+			return record, diag.FromErr(err)
+		}
+		return record, append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Couldn't authenticate to API, please check AWS credentials",
+			Detail:   responseBody["message"],
+		})
+	} else if response.StatusCode == 404 {
+		var responseBody map[string]string
+		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
+			return record, diag.FromErr(err)
+		}
+		return record, append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Couldn't find record with given id",
+			Detail:   responseBody["message"],
+		})
+	} else if response.StatusCode != 200 {
+		// Create error message for any other unexpected errors
+		body, _ := io.ReadAll(response.Body)
+		return record, append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Unexpected error message from API",
+			Detail:   fmt.Sprintf("HTTP %d: %s", response.StatusCode, body),
+		})
+	}
+
+	if err := json.NewDecoder(response.Body).Decode(&record); err != nil {
+		return record, diag.FromErr(err)
+	}
+	return record, diags
+}
+
+func (c *ApiClient) deleteRecord(name string) diag.Diagnostics {
+	var diags diag.Diagnostics
+
+	client := &http.Client{Timeout: 10 * time.Second}
+	request, err := http.NewRequest(http.MethodDelete, fmt.Sprintf("%s/v2/records/%s", HostURL, name), strings.NewReader(""))
+	if err != nil {
+		return diag.FromErr(err)
+	}
+	authorizationHeaders := signRequest(request, c.AccessKeyId, c.SecretAccessKey, c.SessionToken)
+	request.Header.Add("X-Amz-Security-Token", c.SessionToken)
+	request.Header.Add("X-Amz-Date", authorizationHeaders.date)
+	request.Header.Add("Authorization", authorizationHeaders.authorizationHeaders)
+	request.Header.Add("content-type", "application/json")
+	request.Header.Add("x-amz-content-sha256", fmt.Sprintf("%x", authorizationHeaders.payloadHash))
+	request.Header.Set("User-Agent", c.UserAgent)
+
+	response, err := client.Do(request)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+	defer response.Body.Close()
+
+	if response.StatusCode == 403 {
+		// Create proper error message if AWS credentials are not valid, probably because they expired
+		var responseBody map[string]string
+		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
+			return diag.FromErr(err)
+		}
+		return append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Couldn't authenticate to API, please check AWS credentials",
+			Detail:   responseBody["message"],
+		})
+	} else if response.StatusCode == 404 {
+		var responseBody map[string]string
+		if err = json.NewDecoder(response.Body).Decode(&responseBody); err != nil {
+			return diag.FromErr(err)
+		}
+		return append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Couldn't find record with given name",
 			Detail:   responseBody["message"],
 		})
 	} else if response.StatusCode != 204 {