+++ examples/data-sources/csd_zone/data-source.tf
@@ -1,3 +1,3 @@
-data "csd_zone" "myzone" {
+data "csd_zone_delegation" "my_zone_delegation" {
   name = "myzone.idealo.tools"
 }