+++ examples/main.tf
@@ -1,9 +1,10 @@
 terraform {
-  required_version = "~> 1.3"
+  required_version = "~>1.3"
   required_providers {
     csd = {
-      version = "~> 1.0"
+      version = "~>2.0"
       source  = "idealo/csd"
+      #source  = "idealo.com/transport/csd"
     }
     aws = {
       source  = "hashicorp/aws"
@@ -20,45 +21,76 @@ provider "aws" {
 provider "csd" {}
 
 resource "aws_route53_zone" "myzone" {
-  name = "myzone.idealo.tools"
+  name = "myzone2.idealo.tools"
 }
 
-#resource "csd_zone" "myzone" {
-#  name = aws_route53_zone.myzone.name
-#  name_servers = aws_route53_zone.myzone.name_servers
-#}
-#
-#output "test_name" {
-#  value = csd_zone.myzone.name
-#}
-#
-#output "test_name_servers" {
-#  value = csd_zone.myzone.name_servers
+resource "csd_zone_delegation" "myzone" {
+  name         = aws_route53_zone.myzone.name
+  name_servers = aws_route53_zone.myzone.name_servers
+}
+
+resource "aws_route53_record" "myrecord" {
+  zone_id = aws_route53_zone.myzone.id
+  name    = "foo"
+  type    = "TXT"
+  ttl     = 60
+  records = ["bar"]
+}
+
+resource "csd_record" "myrecord" {
+  name   = "_acme-challenge.myrecord.myzone2.idealo.tools"
+  rrtype = "TXT"
+  value  = "foobar"
+}
+
+resource "csd_record" "myrecord2" {
+  name   = "myrecord2.myzone2.idealo.tools"
+  rrtype = "CNAME"
+  value  = "foobar.edgekey.net."
+}
+
+#data "csd_record" "myrecord" {
+#  name = "myrecord.idealo.tools"
 #}
 
-#data "csd_zones" "all" {}
-#
-#output "test_data_read_zones" {
-#  value = data.csd_zones.all
+#output "myrecord" {
+#  value = data.csd_record.myrecord.value
 #}
-#
-#data "csd_zone" "myzone" {
+
+#data "csd_records" "all" {}
+
+#output "all_records" {
+#  value = data.csd_records.all
+#}
+
+#resource "aws_route53_zone" "my_zone" {
 #  name = "myzone.idealo.tools"
 #}
-#
-#output "test_data_read_zone" {
-#  value = data.csd_zone.myzone.name_servers
+
+#resource "csd_zone_delegation" "my_zone_delegation" {
+#  name = aws_route53_zone.my_zone.name
+#  name_servers = aws_route53_zone.my_zone.name_servers
 #}
 
-#resource "csd_zone" "confluence" {
-#  name = "confluence.idealo.tools"
-#  name_servers = [
-#    "ns1.aws.example.net",
-#    "ns2.aws.example.com"
-#  ]
-#  owner = "123456789"
+#output "test_name" {
+#  value = csd_zone_delegation.my_zone_delegation.name
 #}
-#
-#output "test_resource_create_zone" {
-#  value = csd_zone.confluence
+
+#output "test_name_servers" {
+#  value = csd_zone_delegation.my_zone_delegation.name_servers
 #}
+
+#data "csd_zone_delegations" "all" {}
+
+#output "test_data_read_zone_delegations" {
+#  value = data.csd_zone_delegations.all
+#}
+
+#data "csd_zone_delegation" "my_zone_delegation" {
+#  name = "myzone.idealo.tools"
+#}
+
+#output "test_data_read_zone_delegation" {
+#  value = data.csd_zone_delegation.my_zone_delegation.name_servers
+#}
+