+++ examples/resources/csd_zone/resource.tf
@@ -1,8 +1,8 @@
-resource "aws_route53_zone" "myzone" {
+resource "aws_route53_zone" "my_zone" {
   name = "myzone.idealo.tools"
 }
 
-resource "csd_zone" "myzone" {
-  name         = aws_route53_zone.myzone.name
-  name_servers = aws_route53_zone.myzone.name_servers
+resource "csd_zone_delegation" "my_zone_delegation" {
+  name         = aws_route53_zone.my_zone.name
+  name_servers = aws_route53_zone.my_zone.name_servers
 }